Debugging, input and display implementation

+++ server/server/src/led.cpp
@@ -2,7 +2,7 @@
 #include "led.hpp"
 
 // Bind an LED to a provided pin.
-void Led::Bind(int pin) {
+void Led::Bind(uint8_t pin) {
     // Make sure the mode of the bound pin is set correctly.
     pinMode(pin, OUTPUT);
     this->_pin = pin;
@@ -25,7 +25,7 @@ void Led::Toggle() {
     }
 }
 
-void Led::SetState(int state) {
+void Led::SetState(uint8_t state) {
     this->_state = state;
     digitalWrite(this->_pin, state);
 }
